requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
Pillow==8.1.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
